requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
asyncio-throttle==1.0.2
aiohttp==3.9.3 